refactor: formulario para modificar
